PMD fixes and merged MYT-38

+++ src/main/java/uk/gov/hmcts/dts/mytime/services/UserService.java
@@ -1,13 +1,12 @@
 package uk.gov.hmcts.dts.mytime.services;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.dts.mytime.entities.UserEntity;
+import uk.gov.hmcts.dts.mytime.exceptions.NotFoundException;
 import uk.gov.hmcts.dts.mytime.models.UserModel;
 import uk.gov.hmcts.dts.mytime.repository.UserRepo;
 
 @Service
-@Slf4j
 public class UserService {
     private final UserRepo userRepo;
 
@@ -17,23 +16,25 @@ public class UserService {
 
     public UserModel getById(int id) {
 
+        if (!userRepo.existsById(id)) {
+            throw new NotFoundException("No user was found");
+        }
         return new UserModel(userRepo.findById(id));
     }
 
     public void saveUser(UserModel userModel) {
-        log.info("Updating user {}", userModel.toString());
 
         final UserEntity userEntity = new UserEntity(userModel);
 
-        try {
-            userRepo.save(userEntity);
-            log.info("User save successful");
-        } catch (Exception e) {
-            log.info("Error saving {}", e.getMessage());
-        }
+        userRepo.save(userEntity);
+
     }
 
     public void deleteUser(int id) {
+
+        if (!userRepo.existsById(id)) {
+            throw new NotFoundException("No user found to delete");
+        }
 
         userRepo.deleteById(id);
     }
